+++ spec/models.rb
@@ -6,10 +6,10 @@ class Post < ActiveRecord::Base
   belongs_to :user
   belongs_to :category
 
-  redis_search_index(title_field: :title,
-                     score_field: :hits,
-                     condition_fields: [:category_id, :user_id],
-                     ext_fields: [:category_name, :user_name])
+  redis_search title_field: :title,
+               score_field: :hits,
+               condition_fields: [:category_id, :user_id],
+               ext_fields: [:category_name, :user_name]
 
   def category_name
     category.name unless category.blank?
@@ -27,20 +27,17 @@ class User < ActiveRecord::Base
 
   has_many :posts
 
-  redis_search_index(title_field: :name,
-                     alias_field: :alias,
-                     score_field: :score,
-                     condition_fields: [:gender],
-                     prefix_index_enable: true,
-                     ext_fields: [:email])
+  redis_search title_field: :name,
+               alias_field: :alias,
+               score_field: :score,
+               condition_fields: [:gender],
+               ext_fields: [:email]
 end
 
 class Category < ActiveRecord::Base
   include Redis::Search
 
-  redis_search_index(title_field: :name,
-                     prefix_index_enable: true,
-                     ext_fields: [])
+  redis_search title_field: :name, ext_fields: []
 end
 
 class Admin < User
@@ -49,9 +46,8 @@ end
 class Company < ActiveRecord::Base
   include Redis::Search
 
-  redis_search_index(title_field: :name,
-                     prefix_index_enable: true,
-                     class_name: 'Company')
+  redis_search title_field: :name,
+               class_name: 'Company'
 end
 
 class Firm < Company